requirements.txt
pandas>=2.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
dbt-core>=1.7.0
dbt-postgres>=1.7.0
python-dotenv>=1.0.0
